python/requirements.txt
enum34
flask
flask-sqlalchemy
flask-mysqldb
flask-marshmallow